requirements.txt
speechrecognition
gtts
pygame
wikipedia-api
datetime
pyaudio
uuid
